[product] add BIGSERIAL for id col in product table

+++ product/src/main/resources/db/changelog/ddl/changelog-0001.sql
@@ -2,7 +2,7 @@
 
 --changeset samhoho:1
 CREATE TABLE product (
-    id INT NOT NULL,
+    id BIGSERIAL NOT NULL,
     name VARCHAR(255),
     description VARCHAR(255),
     short_description VARCHAR(255),
